nerface_code/requirements_pip.txt
absl-py==0.9.0
blessings==1.7
blinker==1.4
brotlipy==0.7.0
cachetools==3.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
ffmpeg==1.4
freetype-py==2.2.0
google-auth @ file:///home/conda/feedstock_root/build_artifacts/google-auth_1588876379852/work
google-auth-oauthlib==0.4.1
gpustat==0.6.0
grpcio==1.27.2
idna==2.9
imageio==2.8.0
kiwisolver==1.2.0
Markdown==3.2.1
matplotlib==3.3.0
networkx==2.4
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1588605082776/work
nvidia-ml-py3==7.352.0
oauthlib==3.0.1
olefile==0.46
opencv-python-headless==4.4.0.40
Pillow==7.1.2
protobuf==3.11.4
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
pyglet==1.5.7
PyJWT==1.7.1

pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrender==0.1.43
PySocks==1.7.1
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.2.0
rsa==4.0
scipy==1.5.2
six==1.14.0
tensorboard==2.1.1




trimesh==3.7.3
urllib3==1.25.9
Werkzeug==1.0.1

